Refactor to Fix Broken Months
For unknown reasons this line of code is filtering the entire list of rental listings despite the fact we can confirm the bedroom counts in many of the listings in the list is greater than 0. This updated approach does not suffer from this issue, but does intorduce a small discrepency in outputted results for currently working listings.