Fix Incorrect Filtering
After analysis I determined that `listing['bedrooms' >=0]` was not properly filteirng negative bedrooms. A suggestion by Conor lead to this different option which seems to work in the way we expect. This doesn’t answer the question of why the initial code is not working on some of the data, but now that we have consistent results with the new method, it seems to be less of an issue.